perf: Construct datetimes and timedeltas via the C API
Replace Python-level timedelta(...) and datetime.fromtimestamp calls with
PyDelta::new and PyDateTime::from_timestamp. Rust unit tests assert parity
with the Python constructors.

MetadataResponse roundtrip barely exercises datetime construction, so
there is not much effect on the existing benchmark. Add a
ProduceResponse payload with log_append_time on every partition (1000
topics × 12) so rigorous numbers can compare C-API vs Python-constructor
paths on a datetime-heavy workload.

Rigorous roundtrip (MetadataResponse, 1000 topics × 12 partitions):
- vs HEAD^: 9.34 us → 9.37 us (1.00×)
- vs pure Python implementation: 108 us → 9.37 us (11.5×)

Rigorous produce_response (ProduceResponse, 1000 topics × 12 partitions):
- vs HEAD^: 10.7 us → 8.48 us (1.26×)
- vs pure Python implementation: 96.3 us → 8.48 us (11.4×)

+++ rust/src/schema.rs
@@ -205,7 +205,7 @@ fn read_primitive(
         }
         PrimitiveKind::TimedeltaI32 => {
             let (ms, sz) = internal_read_timedelta_i32(bytes)?;
-            let v = instantiate_timedelta(py, ms)?;
+            let v = instantiate_timedelta(py, i64::from(ms))?;
             (v.into_bound(py).into_any(), sz)
         }
         PrimitiveKind::TimedeltaI64 => {
